% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{medianImpute}
\alias{medianImpute}
\title{Imputation by median}
\usage{
medianImpute(learn, val)
}
\arguments{
\item{learn}{The training data.}

\item{val}{The validation data, or \code{NULL}.}
}
\value{
A list of two components:
\item{imputed_learn}{ The imputed training data. }
\item{imputed_val}{ The imputed validation data, or \code{NULL} when 
\code{val = NULL}. }
}
\description{
For each incomplete variable, replace the missing values using the median of
observed values in training data. Note that for factor/integer variables, the
\strong{mode} is used instead of median.
}
