% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputation.R
\name{miceImpute}
\alias{miceImpute}
\title{Imputation by averaging multiple complete data from mice}
\usage{
miceImpute(learn, val, mice.params)
}
\arguments{
\item{learn}{The training data.}

\item{val}{The validation data, or \code{NULL}.}

\item{mice.params}{A list of arguments specified for the \code{mice} function.}
}
\value{
A list of two components:
\item{imputed_learn}{ The imputed training data. }
\item{imputed_val}{ The imputed validation data, or \code{NULL} when 
\code{val = NULL}. }
}
\description{
First, generate \code{m} multivariate imputations by chained equations (MICE) 
using the \code{mice} function from the \pkg{mice} package; then average
over the \code{m} imputations. Note that for factor/integer variables, the
\strong{mode} is used as average instead of mean.
}
