% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.missSuperLearner.R
\name{predict.missSuperLearner}
\alias{predict.missSuperLearner}
\title{Predict method for missSuperLearner object}
\usage{
\method{predict}{missSuperLearner}(object, newdata, X = NULL, Y = NULL, onlySL = FALSE, ...)
}
\arguments{
\item{object}{Fitted object from \code{missSuperLearner}.}

\item{newdata}{New X values for prediction.}

\item{X}{Original data set used to fit \code{object}, if needed.}

\item{Y}{Original outcome used to fit \code{object}, if needed.}

\item{onlySL}{Logical. If \code{TRUE}, only compute predictions for 
algorithms with non-zero coefficients in the super learner object. Default 
is \code{FALSE} (computes predictions for all algorithms in library).}

\item{\dots}{Additional arguments passed to the \code{predict.SL.*}
functions}
}
\value{
A list of two components:
\item{pred}{ Predicted values from the \code{missSuperLearner} fit}
\item{library.predict}{ Predicted values for each algorithm in library}
}
\description{
Obtains predictions on a new data set from a missSuperLearner fit. Need
the original data for imputation, or if one of the library algorithms uses 
the original data in its predict method.
}
\details{
If \code{newdata} is omitted, the predicted values from \code{object} are
returned.  

Each algorithm in the Super Learner library needs to have a
corresponding prediction function with the ``predict.'' prefixed onto the
algorithm name (e.g. \code{predict.SL.glm} for \code{SL.glm}).
}
\seealso{
\code{\link{missSuperLearner}}
}
\author{
Pan Zhao \email{pan.zhao@inria.fr}
}
\keyword{models}
