% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.grf.R
\name{predict.SL.grf}
\alias{predict.SL.grf}
\title{Prediction for SL.grf}
\usage{
\method{predict}{SL.grf}(object, newdata, family, ...)
}
\arguments{
\item{object}{SL.grf object}

\item{newdata}{Dataframe to generate predictions}

\item{family}{Gaussian or binomial}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.grf
}
\seealso{
\code{\link{SL.grf}} \code{\link[grf]{regression_forest}}
\code{\link[grf]{probability_forest}}
\code{\link[grf]{predict.regression_forest}}
\code{\link[grf]{predict.probability_forest}}
}
