% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.grf.R
\name{SL.grf}
\alias{SL.grf}
\title{Wrapper for grf}
\usage{
SL.grf(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{The outcome.}

\item{X}{The covariates.}

\item{newX}{Test dataframe.}

\item{family}{Gaussian or binomial.}

\item{obsWeights}{Observation-level weights.}

\item{...}{Any remaining arguments.}
}
\value{
A trained regression/probability forest object. If 
\code{tune.parameters} is enabled, then tuning information will be included 
through the `tuning.output` attribute.
}
\description{
Wrapper function for generalized random forests algorithm
via \code{regression_forest()} or \code{probability_forest()} from the 
\pkg{grf} package.

Trains a regression forest that can be used to estimate
the conditional mean function \eqn{mu(x) = E[Y | X = x]}, or a probability 
forest that can be used to estimate the conditional class probabilities
\eqn{P[Y = 1 | X = x]}.
}
\seealso{
\code{\link[grf]{regression_forest}}
\code{\link[grf]{probability_forest}}
}
